pyproject.toml
[tool.poetry]
name = "energy-box-control"
version = "0.1.0"
description = ""
authors = ["Boudewijn van Groos <boudewijn.vangroos@foundationzero.org>"]
readme = "README.md"
packages = [{include = "energy_box_control"}]

[tool.poetry.dependencies]
python = "^3.12"
pint = "^0.23"
scipy = "^1.12.0"
paho-mqtt = "^2.0.0"
quart = "^0.19.4"
influxdb-client = { extras = ["async"], version = "^1.41.0" }
fluxy = { git = "https://github.com/foundation-zero/fluxy.git", branch = "main" }
python-dotenv = "^1.0.1"
quart-schema = { extras = ["pydantic"], git = "https://github.com/pgjones/quart-schema.git", branch = "main" }
pandas = "^2.2.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.0.0"
hypothesis = "^6.98.4"
pytest-asyncio = "^0.23.6"
pytest-mock = "^3.12.0"
freezegun = "^1.4.0"

[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
pyright = "^1.1.350"

[tool.pyright]
strict = ["energy_box_control"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
run_power_hub_simulation = 'energy_box_control.simulation:run'
run_power_hub_api = 'energy_box_control.api.api:run'
